yolo/requirements.txt
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
pillow>=10.0.0

